+++ src/store/store.ts
@@ -1,4 +1,3 @@
-
 import { createStore, applyMiddleware, Store } from "redux"
 import thunk from "redux-thunk"
 import { reducer } from "src/context/reducer";